Good_faith_Device_Power.md
Changed a lot, had a very good meeting with the ME team and got to discuss in detail what the motor needed to accomplish and how I needed to power it. Cut the fulfilling constraints section because I looked at some older ones and most people just do in in analysis.